Fix resizing by applying Darrell's patch
I don't get what's going on! Darrell's email:

"Yeah, the GPU buffer objects we’re using for the buffer pool are RBOs, which exist at the OpenGL level and are specific to a particular OpenGL context, but making an OpenGL context current requires a higher-level API (GLX) and object (GLX drawable). Previously, I was reusing the GLX drawable handle from the Pbuffer VirtualGL creates, but that Pbuffer handle will change whenever the window is resized. Thus, a dummy 1x1 Pbuffer is now created in the plugin solely to give us a GLX drawable object to pass to glXMakeContextCurrent(). That dummy Pbuffer isn’t really used. The actual pixels are stored in the RBO, but we can’t use the RBO until a context is active, and the only way to make a context active is to bind it to a GLX drawable. This is a deficiency in GLX and OpenGL, basically, and the 1x1 dummy Pbuffer plugs the gap."